% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_comb.R
\name{bit_comb}
\alias{bit_comb}
\title{Bit combinations}
\usage{
bit_comb(n, boolean = FALSE)
}
\arguments{
\item{n}{\code{numeric} or \code{integer}, number of bits.}

\item{boolean}{\code{boolean}, output a binary \code{matrix} (default) or a boolean \code{matrix}.}
}
\value{
Returns a binary \code{matrix} by default, or a boolean \code{matrix}.
ncol = \code{n}, nrow = \code{2^n}. The bit order in each row is little endian i.e.
least significant bit first.
}
\description{
Generate a matrix with all combinations for an \code{n} bit number.
}
\examples{
bit_comb(3)

}
