% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sets.R
\name{compare_sets}
\alias{compare_sets}
\title{Determine relationships between sets}
\usage{
compare_sets(s, ...)
}
\arguments{
\item{s}{object describing sets to compare. Type can be one of: ...}

\item{...}{}
}
\value{
Returns something...
}
\description{
Determine relationships between sets
}
\examples{
# Example
}
