% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2ellipse.R
\name{df2ellipse}
\alias{df2ellipse}
\title{Generate ellipses from a data frame}
\usage{
df2ellipse(df, n = 360L)
}
\arguments{
\item{df}{\code{data.frame}, each row describes one ellipse. Minimum required info
is 2 columns, \code{x0} and \code{y0} describing the (x, y) coordinates of the centre
of the ellipse.}

\item{n}{\code{integer}, number of points around the ellipse to use. Default is \code{360L}.}
}
\value{
Returns \code{df} with the following additional numeric columns:
\itemize{
\item \code{x} = x coordinate of points describing the ellipse.
\item \code{y} = y coordinate of points describing the ellipse.
\item \code{group} = numeric indicating which points belong to which ellipse.
}
}
\description{
Given a data frame containing sets of geometric
parameters, generate the x,y coordinates of ellipses/super ellipses.

The input data frame can have the following \code{numeric} columns:
\itemize{
\item \code{x0} (required) = x coordinate for ellipse centre.
\item \code{y0} (required) = y coordinate for ellipse centre.
\item \code{a} (optional) = length of semi-major axis. Default is \code{1}.
\item \code{b} (optional) = length of semi-minor axis. Default is \code{1}.
\item \code{angle} (optional) = rotation of the ellipse in radians. Rotation occurs
in a counter-clockwise direction as per the unit circle.
\item \code{m1} (optional) = major semi-diameter of super ellipse. Default is \code{2},
which produces just a normal ellipse.
\item \code{m2} (optional) = minor semi-diameter of super ellipse. Default is to
match \code{m1}.
}
}
\examples{
library(ggplot2)

df <- data.frame(
  x0 = c(0, 1),
  y0 = c(0, 0)
)
df

ellipses <- df2ellipse(df)
head(ellipses)

ggplot(ellipses, aes(x, y, group = group)) +
  geom_polygon(colour = "black", fill = NA) +
  coord_fixed()

}
