% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Generate an ellipse}
\source{
The code for this function is largely copied from \code{ggforce::StatEllip}
which is licensed under the MIT license.
}
\usage{
ellipse(x0 = 0, y0 = 0, a = 1, b = 1, angle = 0, n = 360L, m1 = 2, m2)
}
\arguments{
\item{x0}{\code{numeric}, x coordinate for ellipse centre.}

\item{y0}{\code{numeric}, y coordinate for ellipse centre.}

\item{a}{\code{numeric}, length of semi-major axis.}

\item{b}{\code{numeric}, length of semi-minor axis.}

\item{angle}{\code{numeric}, rotation of ellipse in radians. Rotation occurs
in counter-clockwise direction as per the unit circle.}

\item{n}{\code{integer}, number of points around the ellipse to use.}

\item{m1}{\code{numeric}, major semi-diameter of super ellipse. Default is \code{2},
which produces just a plain ellipse.}

\item{m2}{\code{numeric}, minor semi-diameter of super ellipse. Default is to
match \code{m1}.}
}
\value{
Returns a \code{data.frame} containing coordinates,
with two columns: \code{x} and \code{y}, and \code{n + 1} rows.
}
\description{
Given a set of geometric parameters, generate the x,y coordinates
of an ellipse or super ellipse.
}
\examples{
library(ggplot2)

# default is a unit circle
ggplot(ellipse(), aes(x, y)) +
geom_polygon(colour = "black", fill = NA) +
coord_fixed()

# angle is in radians, and rotation occurs counter-clockwise
ggplot(ellipse(a = 2, angle = pi / 4), aes(x, y)) +
geom_polygon(colour = "black", fill = NA) +
coord_fixed()

# try a super ellipse
ggplot(ellipse(a = 2, angle = pi / 4, m1 = 3), aes(x, y)) +
geom_polygon(colour = "black", fill = NA) +
coord_fixed()

}
