% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvd-package.R
\docType{package}
\name{ggvd-package}
\alias{ggvd}
\alias{ggvd-package}
\title{ggvd: Another Implementation of Venn Diagrams in 'ggplot2'}
\description{
Simple functions to generate 2-4 way Venn diagrams using 
    'ggplot2' with minimal dependencies.
}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com} (\href{https://orcid.org/0000-0002-7151-5971}{ORCID})

}
\keyword{internal}
