% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_venn.R
\name{prepare_venn}
\alias{prepare_venn}
\title{Convert list to geom_venn compatible data.frame}
\usage{
prepare_venn(ls, ...)
}
\arguments{
\item{ls}{Description.}

\item{...}{Description.}
}
\value{
Returns ...
}
\description{
Description.
}
\examples{

lst <- list(
  set1 = c(1:3),
  set2 = c(3:20),
  set3 = c(1, 20:66)
)

prepare_venn(lst, fill = c("blue", "red", "green"))

}
