% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_multi.R
\name{st_multi}
\alias{st_multi}
\alias{st_multi_intersection}
\alias{st_multi_difference}
\alias{st_multi_union}
\title{Perform geometric set intersection, difference, and union with more than two
simple feature geometry collections}
\usage{
st_multi_intersection(..., l = NULL)

st_multi_difference(..., l = NULL)

st_multi_union(..., l = NULL)
}
\arguments{
\item{...}{at least three items are needed if this parameter is used.}

\item{l}{a \code{list} of polygons}
}
\value{
Returns the intersection/union/difference of \code{...}
}
\description{
Perform geometric set intersection, difference, and union with more than two
simple feature geometry collections
}
