% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_venn.R
\name{count_venn}
\alias{count_venn}
\title{Compute counts for Venn Diagram}
\usage{
count_venn(l)
}
\arguments{
\item{l}{Description.}
}
\value{
Returns ...
}
\description{
Description.
}
\examples{

lst <- list(
  set1 = c(1:3),
  set2 = c(3:20),
  set3 = c(1, 20:66)
)

count_venn(lst)

}
