% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_venn.R
\name{count_venn}
\alias{count_venn}
\title{Compute counts for a list of sets}
\usage{
count_venn(l)
}
\arguments{
\item{l}{`list` of up to 26 numeric or character vectors (i.e. sets) to
compare. Within each set there must not be any duplicated elements. If the
list is not named then capital letters starting from A to Z will be
substituted as names.}
}
\value{
Returns a `numeric matrix` with `2^length(l)` rows and
`length(l) + 2` columns. Each row corresponds to the overlap of a particular
combination of sets. The first `length(l)` columns correspond to each set and
contain 1 or 0 indicating membership or not in each set. The second last
column called `count` gives the sum of each row and the last column `percent`
is the the value of `count` for each row, divided by sum of the `count`
column multiplied by 100.
}
\description{
Compute the overlaps of up to 26 sets in a list.
Similar to `vennCounts` from the `limma` package.
}
\examples{

lst <- list(
  set1 = c(1:3),
  set2 = c(3:20),
  set3 = c(1, 20:66)
)

count_venn(lst)

}
