% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_venn.R
\name{geom_venn}
\alias{geom_venn}
\title{Title}
\usage{
geom_venn(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Description.}

\item{data}{Description.}

\item{position}{Description.}

\item{na.rm}{Description.}

\item{show.legend}{Description.}

\item{inherit.aes}{Description.}

\item{...}{Description.}
}
\value{
Description.
}
\description{
Description.
}
