% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_clip.R
\name{poly_clip}
\alias{poly_clip}
\title{Polygon clipping}
\source{
Slightly modified from \code{eulerr::poly_clip()} which is distributed
under the GPLv3 license.
}
\usage{
poly_clip(a, b, op = c("intersection", "union", "minus", "xor"))
}
\arguments{
\item{a, b}{\code{list} or \verb{list of list}, specifying polygons. See details.}

\item{op}{\code{string}, set operation to be performed to combine \code{a} and \code{b}.
One of \code{"intersection"}, \code{"union"}, \code{"minus"}, \code{"xor"} (partially matched).}
}
\value{
Returns a \verb{list of lists}.
}
\description{
A wrapper around \code{\link[polyclip:polyclip]{polyclip::polyclip()}}, which is an interface
to the \emph{Clipper} C++ library.
}
\details{
Each of the arguments \code{a} and \code{b} represents a region in the
Euclidean plane bounded by closed polygons. The format of these arguments
is either:
\itemize{
\item a \code{list} containing two numeric vectors \code{x} and \code{y} giving the
coordinates of the vertices of a single polygon. The last vertex
\strong{should not} repeat the first vertex.
\item a \verb{list of list(x,y)} structures giving the coordinates of the vertices
of several polygons.
}
}
\examples{
#
}
