% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_venn.R
\name{prepare_venn}
\alias{prepare_venn}
\title{Convert list to \code{geom_venn} compatible data.frame}
\usage{
prepare_venn(l, ...)
}
\arguments{
\item{l}{\code{list} of up to 26 numeric or character vectors (i.e. sets) to
compare. Within each set there must not be any duplicated elements. If the
list is not named then capital letters starting from A to Z will be
substituted as names.}

\item{...}{Name-value pairs and other arguments to be passed to
\code{\link[tibble]{add_column}}. See \strong{Examples} section for an example.}
}
\value{
Returns a \code{data.frame} with \code{length(l)} rows and at least three
essential columns: column 1 is \code{set_name} with the names of each set,
column 2 is \code{elements} contains the elements in each set, column 3 is
\code{count} which defines the smallest and the largest set overlap which is
used for the legend limits in \code{geom_venn(type = 'continuous')}. Any additional
columns are added afterward.
}
\description{
Using a list of sets, construct a data.frame with the correct
format for \code{\link{geom_venn}}.
}
\examples{
lst <- list(
  set1 = c(1:3),
  set2 = c(3:20),
  set3 = c(1, 20:66)
)

# Add an additional column called 'fill'
vd <- prepare_venn(lst, fill = c("blue", "red", "green"))

# Plot a Venn diagram
ggplot() +
  geom_venn(aes(set_name = set_name, elements = elements, fill = fill), data = vd) +
  scale_fill_identity()

}
